+++ CHANGES.rst
@@ -13,6 +13,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- #242: Expose geoip database to async tasks.
+
 - Make sure there are no unexpected raven messages left after each test.
 
 - #434: Internal test only changes to test base classes.